header+sidebar+body+footer
This shouldn't be like this. Every file should be saved seperately and we should call them seperately right.
I did this because on the day of submission of this design(saturday as we planned) we can show them that we have done admin page.

so now ill make employee item and srvices page and commit it.